PZ-100 | Open archive in PearlZip through macOs Finder Application + Split out JavaFX MacOS specific code from the main launcher to enable AWT to pick up file requests + ArchiveUtil - Pulled out common code + Mac JavaFX App can handle opening encrypted files (using pre Open dialogs) by double clicking + ErrorAlertConsumer - Removed unnecessary println
